Restructuring BLE services.
