+++ Nemesis-Mod/src/bluetooth/CustomBLEService.cpp
@@ -1,0 +1,7 @@
+#include "CustomBLEService.h"
+
+CustomBLEService::CustomBLEService(BLEUuid uuid) : BLEService(uuid) {
+}
+
+void CustomBLEService::init() {
+}
